# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API Documentation",
        "description": "Swagger/OpenAPI documentation for the Laravel ecommerce backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a single category with its products",
                "operationId": "showCategory",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login an existing user",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get current profile",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "Authenticated user profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update profile",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/change-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change account password",
                "operationId": "changePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout current session",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "MessageResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Logged out successfully"
                }
            },
            "AuthResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Logged in successfully",
                    "user": {
                        "id": 1,
                        "name": "John Doe",
                        "email": "john@example.com",
                        "is_admin": false
                    },
                    "token": "1|abc123token"
                }
            },
            "ProfileResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "user": {
                        "id": 1,
                        "name": "John Doe",
                        "email": "john@example.com",
                        "is_admin": false
                    }
                }
            },
            "LoginRequest": {
                "type": "object",
                "example": {
                    "email": "john@example.com",
                    "password": "secret123"
                }
            },
            "RegisterRequest": {
                "type": "object",
                "example": {
                    "name": "John Doe",
                    "email": "john@example.com",
                    "password": "secret123",
                    "password_confirmation": "secret123"
                }
            },
            "UpdateProfileRequest": {
                "type": "object",
                "example": {
                    "name": "John Doe",
                    "email": "john@example.com"
                }
            },
            "ChangePasswordRequest": {
                "type": "object",
                "example": {
                    "current_password": "oldsecret123",
                    "password": "newsecret123",
                    "password_confirmation": "newsecret123"
                }
            },
            "CategoryResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "data": [
                        {
                            "id": 1,
                            "name": "Vegetables",
                            "slug": "vegetables",
                            "description": "Fresh market vegetables",
                            "products": [
                                {
                                    "id": 10,
                                    "name": "Red Chili",
                                    "slug": "red-chili",
                                    "description": "Hot and fresh chili peppers",
                                    "price": 2.99,
                                    "stock": 50,
                                    "image": "products/chili.jpg"
                                }
                            ]
                        }
                    ]
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use the token returned from login or register in the Authorization header as Bearer <token>.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}